astra*.launch: Allow user to set manager arg
...by specifying only a default rather than an absolute value.

Also move the arg declaration out of the group namespace element so that
it will have the same value everywhere it is used in this launch file.
Otherwise, the camera driver node and the nodelet manager might not agree on
what the name of the manager is, so it would livelock during startup.