astrapro.launch: add external_manager, required argument
Add argument, external_manager, which when true causes the launch file
to not launch a nodelet manager. Instead, it will rely on the user of
the launch file to provide their own nodelet manager.

Add argument, required, which defaults to true. When false, the
nodelet load will not be required="true" (useful for exit_on_load).