Fix eval statement in RGB format
The braces in the eval statement should be parenthesis. This was
causing the value to not evaluate and simply be the string of
the eval statement. The driver was thankfully handling this and
still launching the RGB stream with an appropriate format, which
is why this was missed.

ROS-1843: Orbbec driver depthcam rgb_format default eval
  statement is written incorrectly, not evaluating.