Fix depth-format setting in Gemini 2 launch file
The Y16 format does not appear to be supported by the depthcam driver
(at least given our current configuration/code version), as using this setting
causes it to crash and list the supported configuations. As of the time of
writing, the only supported format (as indicated by the error output)
is RLE (run-length encoding).

Change the default `depth_format` arg to `RLE` to fix the crash.