Daylight the depth_format parameter for gemini2 launches
On Gemini 2 cameras, firmware 1.4.98 requires that this value be set to
"Y16". Prior versions of the firmware required that this value be set to
"RLE". In either case, the driver will not successfully start unless the
correct value has been specified. Therefore, allow calls to roslaunch
to specify this parameter at the top level so that the same launch file
can be used in both cases.