Tweak noise removal filter max size higher
I have noticed with the low mounted gemini 2 that there is
significant noise on obstacles close to the robot. This noise
is on the edge and presents as deep dives past the object.
These pixels were being stuck in the costmap erroneously.

Since these noise blobs are close to the camera they require
the max size to be turned up significantly.

I havent seen any negative impact from turning this up on
other objects in the image. The concern would be potentially
filtering out objects further away in the image, but in those
cases I think the lidars would cover them (and again I see
little difference).