Tune nosie removal filter
On some cameras, I have seen issues with close returns on the body
that prevent movement. Some are due to changes in exposure causing
new areas of the body to be observed. Others are from the returns
on objects at the top edge of the image that seem to waterfall
towards the camera. I am tweaking the noise removal filter to
fix this.

First, change the max size from 240 to 20. The value was increased
to 240 from 80 based upon a faulty description of what this value
does from Orbbec's website. They say that the larger this value,
the more significant the removal. The inverse seems to be the case.
This will allow small blobs to be removed from the image whereas at
a value of 240 they were included.

Turn back down the min diff from 400 to the default of 240. This is
for the waterfalling case. A value of 400 in that scenario was too
large of a threshold for those blobs to be considered noise.