Futher parameretize depth settings, revert changes
Further parameratize the depth work mode and noise filter
max size. This is because the settings to be used by low
and proximity Gemini 2s will be different.

The default settings have the proximity cameras in mind.
With the settings in the include file for the proximity
cameras, this will effectively revert dff06cd and baa3ee1.

After further testing there were multiple issues found with
the changes in those commits. The change in the work mode
for proximity camera introduced a lot more noise on the body,
and the change in the noise removal filter causes valid objects
to not be seen.